requirements.txt
filelock
pillow
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
fire
humanize
requests
tqdm
matplot
scikit-image
scipy
numpy
blobfile
clip @ git+https://github.com/openai/CLIP.git
trimesh

# gradio demo
gradio
